Clean up state pkg's getter and setter (#8533)
