+++ beacon-chain/state/BUILD.bazel
@@ -11,6 +11,7 @@ go_library(
         "setters.go",
         "state_trie.go",
         "types.go",
+        "validator_getters.go",
     ],
     importpath = "github.com/prysmaticlabs/prysm/beacon-chain/state",
     visibility = [
