+++ beacon-chain/state/getters.go
@@ -4,7 +4,6 @@ import (
 	"bytes"
 	"errors"
 	"fmt"
-	"time"
 
 	types "github.com/prysmaticlabs/eth2-types"
 	ethpb "github.com/prysmaticlabs/ethereumapis/eth/v1alpha1"
@@ -13,91 +12,6 @@ import (
 	"github.com/prysmaticlabs/prysm/shared/bytesutil"
 	"github.com/prysmaticlabs/prysm/shared/params"
 )
-
-// EffectiveBalance returns the effective balance of the
-// read only validator.
-func (v ReadOnlyValidator) EffectiveBalance() uint64 {
-	if v.IsNil() {
-		return 0
-	}
-	return v.validator.EffectiveBalance
-}
-
-// ActivationEligibilityEpoch returns the activation eligibility epoch of the
-// read only validator.
-func (v ReadOnlyValidator) ActivationEligibilityEpoch() types.Epoch {
-	if v.IsNil() {
-		return 0
-	}
-	return v.validator.ActivationEligibilityEpoch
-}
-
-// ActivationEpoch returns the activation epoch of the
-// read only validator.
-func (v ReadOnlyValidator) ActivationEpoch() types.Epoch {
-	if v.IsNil() {
-		return 0
-	}
-	return v.validator.ActivationEpoch
-}
-
-// WithdrawableEpoch returns the withdrawable epoch of the
-// read only validator.
-func (v ReadOnlyValidator) WithdrawableEpoch() types.Epoch {
-	if v.IsNil() {
-		return 0
-	}
-	return v.validator.WithdrawableEpoch
-}
-
-// ExitEpoch returns the exit epoch of the
-// read only validator.
-func (v ReadOnlyValidator) ExitEpoch() types.Epoch {
-	if v.IsNil() {
-		return 0
-	}
-	return v.validator.ExitEpoch
-}
-
-// PublicKey returns the public key of the
-// read only validator.
-func (v ReadOnlyValidator) PublicKey() [48]byte {
-	if v.IsNil() {
-		return [48]byte{}
-	}
-	var pubkey [48]byte
-	copy(pubkey[:], v.validator.PublicKey)
-	return pubkey
-}
-
-// WithdrawalCredentials returns the withdrawal credentials of the
-// read only validator.
-func (v ReadOnlyValidator) WithdrawalCredentials() []byte {
-	creds := make([]byte, len(v.validator.WithdrawalCredentials))
-	copy(creds, v.validator.WithdrawalCredentials)
-	return creds
-}
-
-// Slashed returns the read only validator is slashed.
-func (v ReadOnlyValidator) Slashed() bool {
-	if v.IsNil() {
-		return false
-	}
-	return v.validator.Slashed
-}
-
-// CopyValidator returns the copy of the read only validator.
-func (v ReadOnlyValidator) CopyValidator() *ethpb.Validator {
-	if v.IsNil() {
-		return nil
-	}
-	return CopyValidator(v.validator)
-}
-
-// CopyValidator returns the copy of the read only validator.
-func (v ReadOnlyValidator) IsNil() bool {
-	return v.validator == nil
-}
 
 // InnerStateUnsafe returns the pointer value of the underlying
 // beacon state proto object, bypassing immutability. Use with care.
@@ -141,15 +55,15 @@ func (b *BeaconState) CloneInnerState() *pbp2p.BeaconState {
 	}
 }
 
-// HasInnerState detects if the internal reference to the state data structure
+// hasInnerState detects if the internal reference to the state data structure
 // is populated correctly. Returns false if nil.
-func (b *BeaconState) HasInnerState() bool {
+func (b *BeaconState) hasInnerState() bool {
 	return b != nil && b.state != nil
 }
 
 // GenesisTime of the beacon state as a uint64.
 func (b *BeaconState) GenesisTime() uint64 {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return 0
 	}
 
@@ -162,7 +76,7 @@ func (b *BeaconState) GenesisTime() uint64 {
 // genesisTime of the beacon state as a uint64.
 // This assumes that a lock is already held on BeaconState.
 func (b *BeaconState) genesisTime() uint64 {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return 0
 	}
 
@@ -171,7 +85,7 @@ func (b *BeaconState) genesisTime() uint64 {
 
 // GenesisValidatorRoot of the beacon state.
 func (b *BeaconState) GenesisValidatorRoot() []byte {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil
 	}
 	if b.state.GenesisValidatorsRoot == nil {
@@ -187,7 +101,7 @@ func (b *BeaconState) GenesisValidatorRoot() []byte {
 // genesisValidatorRoot of the beacon state.
 // This assumes that a lock is already held on BeaconState.
 func (b *BeaconState) genesisValidatorRoot() []byte {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil
 	}
 	if b.state.GenesisValidatorsRoot == nil {
@@ -199,31 +113,9 @@ func (b *BeaconState) genesisValidatorRoot() []byte {
 	return root
 }
 
-// GenesisUnixTime returns the genesis time as time.Time.
-func (b *BeaconState) GenesisUnixTime() time.Time {
-	if !b.HasInnerState() {
-		return time.Unix(0, 0)
-	}
-
-	b.lock.RLock()
-	defer b.lock.RUnlock()
-
-	return b.genesisUnixTime()
-}
-
-// genesisUnixTime returns the genesis time as time.Time.
-// This assumes that a lock is already held on BeaconState.
-func (b *BeaconState) genesisUnixTime() time.Time {
-	if !b.HasInnerState() {
-		return time.Unix(0, 0)
-	}
-
-	return time.Unix(int64(b.state.GenesisTime), 0)
-}
-
 // Slot of the current beacon chain state.
 func (b *BeaconState) Slot() types.Slot {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return 0
 	}
 
@@ -236,7 +128,7 @@ func (b *BeaconState) Slot() types.Slot {
 // slot of the current beacon chain state.
 // This assumes that a lock is already held on BeaconState.
 func (b *BeaconState) slot() types.Slot {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return 0
 	}
 
@@ -245,7 +137,7 @@ func (b *BeaconState) slot() types.Slot {
 
 // Fork version of the beacon chain.
 func (b *BeaconState) Fork() *pbp2p.Fork {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil
 	}
 	if b.state.Fork == nil {
@@ -261,7 +153,7 @@ func (b *BeaconState) Fork() *pbp2p.Fork {
 // fork version of the beacon chain.
 // This assumes that a lock is already held on BeaconState.
 func (b *BeaconState) fork() *pbp2p.Fork {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil
 	}
 	if b.state.Fork == nil {
@@ -281,7 +173,7 @@ func (b *BeaconState) fork() *pbp2p.Fork {
 
 // LatestBlockHeader stored within the beacon state.
 func (b *BeaconState) LatestBlockHeader() *ethpb.BeaconBlockHeader {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil
 	}
 	if b.state.LatestBlockHeader == nil {
@@ -297,7 +189,7 @@ func (b *BeaconState) LatestBlockHeader() *ethpb.BeaconBlockHeader {
 // latestBlockHeader stored within the beacon state.
 // This assumes that a lock is already held on BeaconState.
 func (b *BeaconState) latestBlockHeader() *ethpb.BeaconBlockHeader {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil
 	}
 	if b.state.LatestBlockHeader == nil {
@@ -322,33 +214,9 @@ func (b *BeaconState) latestBlockHeader() *ethpb.BeaconBlockHeader {
 	return hdr
 }
 
-// ParentRoot is a convenience method to access state.LatestBlockRoot.ParentRoot.
-func (b *BeaconState) ParentRoot() [32]byte {
-	if !b.HasInnerState() {
-		return [32]byte{}
-	}
-
-	b.lock.RLock()
-	defer b.lock.RUnlock()
-
-	return b.parentRoot()
-}
-
-// parentRoot is a convenience method to access state.LatestBlockRoot.ParentRoot.
-// This assumes that a lock is already held on BeaconState.
-func (b *BeaconState) parentRoot() [32]byte {
-	if !b.HasInnerState() {
-		return [32]byte{}
-	}
-
-	parentRoot := [32]byte{}
-	copy(parentRoot[:], b.state.LatestBlockHeader.ParentRoot)
-	return parentRoot
-}
-
 // BlockRoots kept track of in the beacon state.
 func (b *BeaconState) BlockRoots() [][]byte {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil
 	}
 	if b.state.BlockRoots == nil {
@@ -364,7 +232,7 @@ func (b *BeaconState) BlockRoots() [][]byte {
 // blockRoots kept track of in the beacon state.
 // This assumes that a lock is already held on BeaconState.
 func (b *BeaconState) blockRoots() [][]byte {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil
 	}
 	return b.safeCopy2DByteSlice(b.state.BlockRoots)
@@ -373,7 +241,7 @@ func (b *BeaconState) blockRoots() [][]byte {
 // BlockRootAtIndex retrieves a specific block root based on an
 // input index value.
 func (b *BeaconState) BlockRootAtIndex(idx uint64) ([]byte, error) {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil, ErrNilInnerState
 	}
 	if b.state.BlockRoots == nil {
@@ -390,7 +258,7 @@ func (b *BeaconState) BlockRootAtIndex(idx uint64) ([]byte, error) {
 // input index value.
 // This assumes that a lock is already held on BeaconState.
 func (b *BeaconState) blockRootAtIndex(idx uint64) ([]byte, error) {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil, ErrNilInnerState
 	}
 	return b.safeCopyBytesAtIndex(b.state.BlockRoots, idx)
@@ -398,7 +266,7 @@ func (b *BeaconState) blockRootAtIndex(idx uint64) ([]byte, error) {
 
 // StateRoots kept track of in the beacon state.
 func (b *BeaconState) StateRoots() [][]byte {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil
 	}
 	if b.state.StateRoots == nil {
@@ -414,7 +282,7 @@ func (b *BeaconState) StateRoots() [][]byte {
 // StateRoots kept track of in the beacon state.
 // This assumes that a lock is already held on BeaconState.
 func (b *BeaconState) stateRoots() [][]byte {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil
 	}
 	return b.safeCopy2DByteSlice(b.state.StateRoots)
@@ -423,7 +291,7 @@ func (b *BeaconState) stateRoots() [][]byte {
 // StateRootAtIndex retrieves a specific state root based on an
 // input index value.
 func (b *BeaconState) StateRootAtIndex(idx uint64) ([]byte, error) {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil, ErrNilInnerState
 	}
 	if b.state.StateRoots == nil {
@@ -440,7 +308,7 @@ func (b *BeaconState) StateRootAtIndex(idx uint64) ([]byte, error) {
 // input index value.
 // This assumes that a lock is already held on BeaconState.
 func (b *BeaconState) stateRootAtIndex(idx uint64) ([]byte, error) {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil, ErrNilInnerState
 	}
 	return b.safeCopyBytesAtIndex(b.state.StateRoots, idx)
@@ -448,7 +316,7 @@ func (b *BeaconState) stateRootAtIndex(idx uint64) ([]byte, error) {
 
 // HistoricalRoots based on epochs stored in the beacon state.
 func (b *BeaconState) HistoricalRoots() [][]byte {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil
 	}
 	if b.state.HistoricalRoots == nil {
@@ -464,7 +332,7 @@ func (b *BeaconState) HistoricalRoots() [][]byte {
 // historicalRoots based on epochs stored in the beacon state.
 // This assumes that a lock is already held on BeaconState.
 func (b *BeaconState) historicalRoots() [][]byte {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil
 	}
 	return b.safeCopy2DByteSlice(b.state.HistoricalRoots)
@@ -472,7 +340,7 @@ func (b *BeaconState) historicalRoots() [][]byte {
 
 // Eth1Data corresponding to the proof-of-work chain information stored in the beacon state.
 func (b *BeaconState) Eth1Data() *ethpb.Eth1Data {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil
 	}
 	if b.state.Eth1Data == nil {
@@ -488,7 +356,7 @@ func (b *BeaconState) Eth1Data() *ethpb.Eth1Data {
 // eth1Data corresponding to the proof-of-work chain information stored in the beacon state.
 // This assumes that a lock is already held on BeaconState.
 func (b *BeaconState) eth1Data() *ethpb.Eth1Data {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil
 	}
 	if b.state.Eth1Data == nil {
@@ -501,7 +369,7 @@ func (b *BeaconState) eth1Data() *ethpb.Eth1Data {
 // Eth1DataVotes corresponds to votes from eth2 on the canonical proof-of-work chain
 // data retrieved from eth1.
 func (b *BeaconState) Eth1DataVotes() []*ethpb.Eth1Data {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil
 	}
 	if b.state.Eth1DataVotes == nil {
@@ -518,7 +386,7 @@ func (b *BeaconState) Eth1DataVotes() []*ethpb.Eth1Data {
 // data retrieved from eth1.
 // This assumes that a lock is already held on BeaconState.
 func (b *BeaconState) eth1DataVotes() []*ethpb.Eth1Data {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil
 	}
 	if b.state.Eth1DataVotes == nil {
@@ -535,7 +403,7 @@ func (b *BeaconState) eth1DataVotes() []*ethpb.Eth1Data {
 // Eth1DepositIndex corresponds to the index of the deposit made to the
 // validator deposit contract at the time of this state's eth1 data.
 func (b *BeaconState) Eth1DepositIndex() uint64 {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return 0
 	}
 
@@ -549,7 +417,7 @@ func (b *BeaconState) Eth1DepositIndex() uint64 {
 // validator deposit contract at the time of this state's eth1 data.
 // This assumes that a lock is already held on BeaconState.
 func (b *BeaconState) eth1DepositIndex() uint64 {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return 0
 	}
 
@@ -558,7 +426,7 @@ func (b *BeaconState) eth1DepositIndex() uint64 {
 
 // Validators participating in consensus on the beacon chain.
 func (b *BeaconState) Validators() []*ethpb.Validator {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil
 	}
 	if b.state.Validators == nil {
@@ -574,7 +442,7 @@ func (b *BeaconState) Validators() []*ethpb.Validator {
 // validators participating in consensus on the beacon chain.
 // This assumes that a lock is already held on BeaconState.
 func (b *BeaconState) validators() []*ethpb.Validator {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil
 	}
 	if b.state.Validators == nil {
@@ -596,7 +464,7 @@ func (b *BeaconState) validators() []*ethpb.Validator {
 // This assumes that a lock is already held on BeaconState. This does not
 // copy fully and instead just copies the reference.
 func (b *BeaconState) validatorsReferences() []*ethpb.Validator {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil
 	}
 	if b.state.Validators == nil {
@@ -617,7 +485,7 @@ func (b *BeaconState) validatorsReferences() []*ethpb.Validator {
 
 // ValidatorAtIndex is the validator at the provided index.
 func (b *BeaconState) ValidatorAtIndex(idx types.ValidatorIndex) (*ethpb.Validator, error) {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil, ErrNilInnerState
 	}
 	if b.state.Validators == nil {
@@ -637,7 +505,7 @@ func (b *BeaconState) ValidatorAtIndex(idx types.ValidatorIndex) (*ethpb.Validat
 // ValidatorAtIndexReadOnly is the validator at the provided index. This method
 // doesn't clone the validator.
 func (b *BeaconState) ValidatorAtIndexReadOnly(idx types.ValidatorIndex) (ReadOnlyValidator, error) {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return ReadOnlyValidator{}, ErrNilInnerState
 	}
 	if b.state.Validators == nil {
@@ -677,7 +545,7 @@ func (b *BeaconState) validatorIndexMap() map[[48]byte]types.ValidatorIndex {
 // PubkeyAtIndex returns the pubkey at the given
 // validator index.
 func (b *BeaconState) PubkeyAtIndex(idx types.ValidatorIndex) [48]byte {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return [48]byte{}
 	}
 	if uint64(idx) >= uint64(len(b.state.Validators)) {
@@ -694,7 +562,7 @@ func (b *BeaconState) PubkeyAtIndex(idx types.ValidatorIndex) [48]byte {
 
 // NumValidators returns the size of the validator registry.
 func (b *BeaconState) NumValidators() int {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return 0
 	}
 	b.lock.RLock()
@@ -706,7 +574,7 @@ func (b *BeaconState) NumValidators() int {
 // ReadFromEveryValidator reads values from every validator and applies it to the provided function.
 // Warning: This method is potentially unsafe, as it exposes the actual validator registry.
 func (b *BeaconState) ReadFromEveryValidator(f func(idx int, val ReadOnlyValidator) error) error {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return ErrNilInnerState
 	}
 	if b.state.Validators == nil {
@@ -727,7 +595,7 @@ func (b *BeaconState) ReadFromEveryValidator(f func(idx int, val ReadOnlyValidat
 
 // Balances of validators participating in consensus on the beacon chain.
 func (b *BeaconState) Balances() []uint64 {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil
 	}
 	if b.state.Balances == nil {
@@ -743,7 +611,7 @@ func (b *BeaconState) Balances() []uint64 {
 // balances of validators participating in consensus on the beacon chain.
 // This assumes that a lock is already held on BeaconState.
 func (b *BeaconState) balances() []uint64 {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil
 	}
 	if b.state.Balances == nil {
@@ -757,7 +625,7 @@ func (b *BeaconState) balances() []uint64 {
 
 // BalanceAtIndex of validator with the provided index.
 func (b *BeaconState) BalanceAtIndex(idx types.ValidatorIndex) (uint64, error) {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return 0, ErrNilInnerState
 	}
 	if b.state.Balances == nil {
@@ -775,7 +643,7 @@ func (b *BeaconState) BalanceAtIndex(idx types.ValidatorIndex) (uint64, error) {
 
 // BalancesLength returns the length of the balances slice.
 func (b *BeaconState) BalancesLength() int {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return 0
 	}
 	if b.state.Balances == nil {
@@ -791,7 +659,7 @@ func (b *BeaconState) BalancesLength() int {
 // balancesLength returns the length of the balances slice.
 // This assumes that a lock is already held on BeaconState.
 func (b *BeaconState) balancesLength() int {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return 0
 	}
 	if b.state.Balances == nil {
@@ -803,7 +671,7 @@ func (b *BeaconState) balancesLength() int {
 
 // RandaoMixes of block proposers on the beacon chain.
 func (b *BeaconState) RandaoMixes() [][]byte {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil
 	}
 	if b.state.RandaoMixes == nil {
@@ -819,7 +687,7 @@ func (b *BeaconState) RandaoMixes() [][]byte {
 // randaoMixes of block proposers on the beacon chain.
 // This assumes that a lock is already held on BeaconState.
 func (b *BeaconState) randaoMixes() [][]byte {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil
 	}
 
@@ -829,7 +697,7 @@ func (b *BeaconState) randaoMixes() [][]byte {
 // RandaoMixAtIndex retrieves a specific block root based on an
 // input index value.
 func (b *BeaconState) RandaoMixAtIndex(idx uint64) ([]byte, error) {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil, ErrNilInnerState
 	}
 	if b.state.RandaoMixes == nil {
@@ -846,7 +714,7 @@ func (b *BeaconState) RandaoMixAtIndex(idx uint64) ([]byte, error) {
 // input index value.
 // This assumes that a lock is already held on BeaconState.
 func (b *BeaconState) randaoMixAtIndex(idx uint64) ([]byte, error) {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil, ErrNilInnerState
 	}
 
@@ -855,7 +723,7 @@ func (b *BeaconState) randaoMixAtIndex(idx uint64) ([]byte, error) {
 
 // RandaoMixesLength returns the length of the randao mixes slice.
 func (b *BeaconState) RandaoMixesLength() int {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return 0
 	}
 	if b.state.RandaoMixes == nil {
@@ -871,7 +739,7 @@ func (b *BeaconState) RandaoMixesLength() int {
 // randaoMixesLength returns the length of the randao mixes slice.
 // This assumes that a lock is already held on BeaconState.
 func (b *BeaconState) randaoMixesLength() int {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return 0
 	}
 	if b.state.RandaoMixes == nil {
@@ -883,7 +751,7 @@ func (b *BeaconState) randaoMixesLength() int {
 
 // Slashings of validators on the beacon chain.
 func (b *BeaconState) Slashings() []uint64 {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil
 	}
 	if b.state.Slashings == nil {
@@ -899,7 +767,7 @@ func (b *BeaconState) Slashings() []uint64 {
 // slashings of validators on the beacon chain.
 // This assumes that a lock is already held on BeaconState.
 func (b *BeaconState) slashings() []uint64 {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil
 	}
 	if b.state.Slashings == nil {
@@ -913,7 +781,7 @@ func (b *BeaconState) slashings() []uint64 {
 
 // PreviousEpochAttestations corresponding to blocks on the beacon chain.
 func (b *BeaconState) PreviousEpochAttestations() []*pbp2p.PendingAttestation {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil
 	}
 	if b.state.PreviousEpochAttestations == nil {
@@ -929,7 +797,7 @@ func (b *BeaconState) PreviousEpochAttestations() []*pbp2p.PendingAttestation {
 // previousEpochAttestations corresponding to blocks on the beacon chain.
 // This assumes that a lock is already held on BeaconState.
 func (b *BeaconState) previousEpochAttestations() []*pbp2p.PendingAttestation {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil
 	}
 
@@ -938,7 +806,7 @@ func (b *BeaconState) previousEpochAttestations() []*pbp2p.PendingAttestation {
 
 // CurrentEpochAttestations corresponding to blocks on the beacon chain.
 func (b *BeaconState) CurrentEpochAttestations() []*pbp2p.PendingAttestation {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil
 	}
 	if b.state.CurrentEpochAttestations == nil {
@@ -954,7 +822,7 @@ func (b *BeaconState) CurrentEpochAttestations() []*pbp2p.PendingAttestation {
 // currentEpochAttestations corresponding to blocks on the beacon chain.
 // This assumes that a lock is already held on BeaconState.
 func (b *BeaconState) currentEpochAttestations() []*pbp2p.PendingAttestation {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil
 	}
 
@@ -963,7 +831,7 @@ func (b *BeaconState) currentEpochAttestations() []*pbp2p.PendingAttestation {
 
 // JustificationBits marking which epochs have been justified in the beacon chain.
 func (b *BeaconState) JustificationBits() bitfield.Bitvector4 {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil
 	}
 	if b.state.JustificationBits == nil {
@@ -979,7 +847,7 @@ func (b *BeaconState) JustificationBits() bitfield.Bitvector4 {
 // justificationBits marking which epochs have been justified in the beacon chain.
 // This assumes that a lock is already held on BeaconState.
 func (b *BeaconState) justificationBits() bitfield.Bitvector4 {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil
 	}
 	if b.state.JustificationBits == nil {
@@ -993,7 +861,7 @@ func (b *BeaconState) justificationBits() bitfield.Bitvector4 {
 
 // PreviousJustifiedCheckpoint denoting an epoch and block root.
 func (b *BeaconState) PreviousJustifiedCheckpoint() *ethpb.Checkpoint {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil
 	}
 	if b.state.PreviousJustifiedCheckpoint == nil {
@@ -1009,7 +877,7 @@ func (b *BeaconState) PreviousJustifiedCheckpoint() *ethpb.Checkpoint {
 // previousJustifiedCheckpoint denoting an epoch and block root.
 // This assumes that a lock is already held on BeaconState.
 func (b *BeaconState) previousJustifiedCheckpoint() *ethpb.Checkpoint {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil
 	}
 
@@ -1018,7 +886,7 @@ func (b *BeaconState) previousJustifiedCheckpoint() *ethpb.Checkpoint {
 
 // CurrentJustifiedCheckpoint denoting an epoch and block root.
 func (b *BeaconState) CurrentJustifiedCheckpoint() *ethpb.Checkpoint {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil
 	}
 	if b.state.CurrentJustifiedCheckpoint == nil {
@@ -1034,7 +902,7 @@ func (b *BeaconState) CurrentJustifiedCheckpoint() *ethpb.Checkpoint {
 // currentJustifiedCheckpoint denoting an epoch and block root.
 // This assumes that a lock is already held on BeaconState.
 func (b *BeaconState) currentJustifiedCheckpoint() *ethpb.Checkpoint {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil
 	}
 
@@ -1044,7 +912,7 @@ func (b *BeaconState) currentJustifiedCheckpoint() *ethpb.Checkpoint {
 // MatchCurrentJustifiedCheckpoint returns true if input justified checkpoint matches
 // the current justified checkpoint in state.
 func (b *BeaconState) MatchCurrentJustifiedCheckpoint(c *ethpb.Checkpoint) bool {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return false
 	}
 	if b.state.CurrentJustifiedCheckpoint == nil {
@@ -1060,7 +928,7 @@ func (b *BeaconState) MatchCurrentJustifiedCheckpoint(c *ethpb.Checkpoint) bool 
 // MatchPreviousJustifiedCheckpoint returns true if the input justified checkpoint matches
 // the previous justified checkpoint in state.
 func (b *BeaconState) MatchPreviousJustifiedCheckpoint(c *ethpb.Checkpoint) bool {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return false
 	}
 	if b.state.PreviousJustifiedCheckpoint == nil {
@@ -1075,7 +943,7 @@ func (b *BeaconState) MatchPreviousJustifiedCheckpoint(c *ethpb.Checkpoint) bool
 
 // FinalizedCheckpoint denoting an epoch and block root.
 func (b *BeaconState) FinalizedCheckpoint() *ethpb.Checkpoint {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil
 	}
 	if b.state.FinalizedCheckpoint == nil {
@@ -1091,7 +959,7 @@ func (b *BeaconState) FinalizedCheckpoint() *ethpb.Checkpoint {
 // finalizedCheckpoint denoting an epoch and block root.
 // This assumes that a lock is already held on BeaconState.
 func (b *BeaconState) finalizedCheckpoint() *ethpb.Checkpoint {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return nil
 	}
 
@@ -1100,7 +968,7 @@ func (b *BeaconState) finalizedCheckpoint() *ethpb.Checkpoint {
 
 // FinalizedCheckpointEpoch returns the epoch value of the finalized checkpoint.
 func (b *BeaconState) FinalizedCheckpointEpoch() types.Epoch {
-	if !b.HasInnerState() {
+	if !b.hasInnerState() {
 		return 0
 	}
 	if b.state.FinalizedCheckpoint == nil {
